# bestbank_be/docs/swagger.yml
openapi: 3.0.0

info:
  title: Banking App API
  version: 1.0.0

servers:
  - url: http://localhost:3000

paths:

  /api/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, phone]
              properties:
                email:
                  type: string
                password:
                  type: string
                phone:
                  type: string
      responses:
        '201':
          description: User created, OTP sent
        '400':
          description: Invalid input or email already exists

  /api/auth/verify-otp:
    post:
      summary: Verify OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, otp]
              properties:
                email:
                  type: string
                otp:
                  type: string
      responses:
        '200':
          description: Verified, returns JWT token
        '400':
          description: Wrong OTP

  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Returns JWT token
        '401':
          description: Invalid credentials

  /api/auth/logout:
    post:
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out

  /api/account/me:
    get:
      summary: Get balance and profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Returns email, phone, balance
        '401':
          description: Unauthorized

  /api/transactions:
    get:
      summary: Get transaction history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Returns list of transactions
        '401':
          description: Unauthorized
    post:
      summary: Transfer money to another user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [receiverEmail, amount]
              properties:
                receiverEmail:
                  type: string
                amount:
                  type: number
      responses:
        '200':
          description: Transfer successful
        '400':
          description: Insufficient funds or user not found
        '401':
          description: Unauthorized

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT